feat: Add Dockerfile and project dependencies
This commit adds a Dockerfile that sets up a Python 3.9-slim environment for running the project. It also includes the installation of project dependencies specified in requirements.txt. The project directory and the .env file are copied to the container as well. Finally, the main.py script is executed using the "python3" command.